Notebooks for my graphs and games class, initial commit.
All of the notebooks are in place, but they all need work still.

outline has the schedule for the 5 class sessions and a few bullet points on what to do for some of the classes. There are also several TODOs there to prepare the remaining materials.

intro_graphs_python has a demo of numbers, strings, and a bit of lists; still need variables, functions, more on lists, while, if/else, and more on graphs (some is there).

bfs_dfs is has implementations and nice visualizations (either while searching or creating a video afterwards), including plotting the path on the maze. One can generate random "mazes". A bit more is needed in terms of explaining the algorithms.

dijkstra_a_star has a function for plotting terrain but still needs implementations and explanations for both algorithms. There's a link to the website I made for creating a terrain and a function that lets you paste that and returns a numpy array.

personal_projects has the 8-puzzle with classes and a BFS solver. A visualization to show following the solved path is in progress (but probably won't show up in this repo; it'll be on the website with the Create Map webpage). Ideas for other projects are still needed as well as an outline of what to say when introducing the projects.

advanced is empty. Ideally some jump point search implementation and visualization would go here.

utils has the node and graph classes that I made for Adam. I haven't used them for this class, as we've just been using grids, but I kept them in case they might be useful as examples. More functions will probably be moved from the notebooks to utils later.