Created Terrain class to make things nicer, especially with plotting.
Terrain class has methods for saving/loading a map, for plotting the terrain, and for plotting paths on it. It should work even if the difficulties of the terrain types aren't in the same order as the terrain types are in the list (i.e. not in the order road, grassy, forest, rocky), though I haven't tested this yet.

Also created directories for movies and images to keep things cleaner.