15 puzzle now supports any size of puzzle; general bfs_dfs and get_neighbors for grids functions in graph_search.
fifteen_puzzle.py has a new implementation using numpy arrays that makes it easy to have a puzzle of any size.

graph_search.py has a function based on the bfs_dfs one in bfs_dfs.ipynb. The plotting/video-making functionality is removed, and it's more general. One can pass in the functions to use to get neighbors and to check if a node is a goal state, so it should work very broadly. In case the nodes aren't hashable, it even supports a function converting from nodes to hashable versions of them (note that there will be some extra overhead from calling an identity function on nodes that are already hashable).